Resolved this not working - Parse function error
The parser was slicing 1 extra character off everything! So if it got the line `event: put`, instead of `event` and `put`, it would cut it to `even` and `pu`.
This commit fixes that issue, as well as convert the byte buffer into a string just once when the line is added to the list of lines to check. This lets us remove the `EVENT` and `DATA` constants, and not go through the hassle of calling the `.get_string_from` functions multiple times.